+++ modules/azure-app-gateway/outputs.tf
@@ -1,3 +1,4 @@
 output "id" {
-    value = azurerm_application_gateway.application_gateway.id
+  description = "Resource ID of the Application Gateway."
+  value       = azurerm_application_gateway.application_gateway.id
 }