+++ modules/azure-disks-backup/data.tf
@@ -1,9 +1,9 @@
-# https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/data-sources/resource_group
+# https://registry.terraform.io/providers/hashicorp/azurerm/4.5.0/docs/data-sources/resource_group
 data "azurerm_resource_group" "this" {
-    name = var.resource_group_name
+  name = var.resource_group_name
 }
 
-# https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/data-sources/managed_disk
+# https://registry.terraform.io/providers/hashicorp/azurerm/4.5.0/docs/data-sources/managed_disk
 data "azurerm_managed_disk" "this" {
   for_each            = { for instance in var.backup_instances : instance.disk_name => instance }
   name                = each.value.disk_name